Refactor Chef::Mixin::WideString to remove implicit Windows dependency.
