+++ lib/chef/win32/api/file.rb
@@ -20,6 +20,7 @@
 require 'chef/win32/api'
 require 'chef/win32/api/security'
 require 'chef/win32/api/system'
+require 'chef/win32/unicode'
 
 class Chef
   module ReservedNames::Win32
