+++ lib/chef/win32/api/net.rb
@@ -17,6 +17,7 @@
 #
 
 require 'chef/win32/api'
+require 'chef/win32/unicode'
 
 class Chef
   module ReservedNames::Win32
