+++ lib/chef/win32/crypto.rb
@@ -19,6 +19,7 @@
 require 'chef/win32/error'
 require 'chef/win32/api/memory'
 require 'chef/win32/api/crypto'
+require 'chef/win32/unicode'
 require 'digest'
 
 class Chef
