+++ lib/chef/win32/net.rb
@@ -18,7 +18,7 @@
 
 require 'chef/win32/api/net'
 require 'chef/win32/error'
-require 'chef/mixin/wstring'
+require 'chef/mixin/wide_string'
 
 class Chef
   module ReservedNames::Win32
